+++ R/opq.R
@@ -225,16 +225,6 @@ opq_osm_id <- function (id = NULL, type = NULL, open_url = FALSE)
     opq
 }
 
-#' @rdname add_osm_feature
-#' @export
-add_feature <- function (opq, key, value, key_exact = TRUE,
-                             value_exact = TRUE, match_case = TRUE, bbox = NULL)
-{
-    message ('add_feature() is deprecated; please use add_osm_feature()')
-    add_osm_feature (opq, key, value, key_exact = TRUE,
-                     value_exact = TRUE, match_case = TRUE, bbox = NULL)
-}
-
 #' Convert an overpass query into a text string
 #'
 #' Convert an osmdata query of class opq to a character string query to